added changes i made many days ago - not entirely sure what they are...

+++ src/cnmgrStructs.h
@@ -41,9 +41,9 @@ struct cnmgrCameraNode
 	float position[3];
 	float rotation[4][4];
 	float projection[4][4];
-	GLdouble fov;
-	GLdouble zNear;
-	GLdouble zFar;
+	float fov;
+	float zNear;
+	float zFar;
 	void *parent;
 	int nextType;
 	void *next;
